heroku second commit

+++ client/js/controllers/PostController.js
@@ -12,22 +12,27 @@ app.controller(`PostController`, [`$scope`, `PostsService`, `$routeParams`, `$lo
         })
 
     $scope.submitEditPost = function(post) {
-      console.log(`edit post controller is firing`);
+      //if cookies.id = post.users_id, then give them the error
+        console.log(`edit post controller is firing`);
         const editedPost = $scope.post
         PostsService.putOnePost(editedPost)
             .then(function() {
-              console.log(`this is edited post`, editedPost)
-                $location.url('/')
+                console.log(`this is edited post`, editedPost)
+                $location.url(`/`)
             })
     }
 
-      $scope.deletePost = function(post) {
-        console.log(`delete controller is firing`);
-        const id = post.id
-        PostsService.deleteOnePost(id)
-        .then(function() {
-          $location.url(`/`)
-        })
-      }
+    $scope.deletePost = function(post) {
+        let check = confirm(`Are you sure you want to delete the post?`)
+        if (check) {
+            const id = post.id
+            PostsService.deleteOnePost(id)
+                .then(function() {
+                    $location.url(`/`)
+                })
+        } else {
+            $location.url(`#/{{post.id}}`)
+        }
+    }
 
 }])
